+++ src/main/java/com/qpolla/entry/data/entity/EntryEntity.java
