+++ src/main/java/com/qpolla/poll/data/EnumPollCategory.java
@@ -1,4 +1,4 @@
-package com.qpolla.data.entity.poll;
+package com.qpolla.poll.data;
 
 import lombok.Getter;
