+++ src/main/java/com/qpolla/poll/data/entity/PollEntity.java
