+++ src/main/java/com/qpolla/poll/repository/OptionRepository.java
@@ -1,6 +1,6 @@
-package com.qpolla.repository;
+package com.qpolla.poll.repository;
 
-import com.qpolla.data.entity.OptionEntity;
+import com.qpolla.poll.data.OptionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OptionRepository extends JpaRepository<OptionEntity, Long> {}
