+++ src/main/java/com/qpolla/poll/repository/PollRepository.java
@@ -1,6 +1,6 @@
-package com.qpolla.repository;
+package com.qpolla.poll.repository;
 
-import com.qpolla.data.entity.poll.PollEntity;
+import com.qpolla.poll.data.PollEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PollRepository extends JpaRepository<PollEntity, Long> {}
