+++ src/main/java/com/qpolla/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.qpolla.repository;
+package com.qpolla.user.repository;
 
-import com.qpolla.data.entity.user.UserEntity;
+import com.qpolla.user.data.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<UserEntity, Long> {}
