+++ src/test/java/com/qpolla/QPollaApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class GundemioylaApplicationTests {
+class QPollaApplicationTests {
 
 	@Test
 	void contextLoads() {
